Refactoring, imports, docs, other cleaning.
